Refine the structure
